fix: remove legacy code
