+++ .github/workflows/no_std.yaml
@@ -16,7 +16,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        crate: ["superchain", "superchain-primitives", "superchain-registry"]
+        crate: ["superchain"]
     steps:
       - uses: actions/checkout@v4
       - uses: dtolnay/rust-toolchain@stable
