+++ Justfile
@@ -69,14 +69,6 @@ test-features *args='':
 test-docs:
   cargo test --doc --all --locked
 
-# Release the primitives crate
-release-primitives:
-  cargo release publish --package superchain-primitives --execute --no-confirm
-
-# Release the registry crate
-release-registry:
-  cargo release publish --package superchain-registry --execute --no-confirm
-
 # Release the superchain crate
 release-superchain:
   cargo release publish --package superchain --execute --no-confirm
