--- crates/primitives/Cargo.toml
@@ -1,39 +1,0 @@
-[package]
-name = "superchain-primitives"
-description = "Primitive Types for the Superchain Registry"
-version.workspace = true
-edition.workspace = true
-license.workspace = true
-authors.workspace = true
-repository.workspace = true
-
-[dependencies]
-# Alloy
-alloy-sol-types.workspace = true
-alloy-primitives.workspace = true
-alloy-genesis.workspace = true
-alloy-consensus.workspace = true
-alloy-eips.workspace = true
-
-# Misc
-anyhow.workspace = true
-
-# `serde` feature dependencies
-serde = { workspace = true, optional = true }
-serde_repr = { workspace = true, optional = true }
-alloy-serde = { workspace = true, optional = true }
-
-# `arbitrary` feature dependencies
-arbitrary = { workspace = true, optional = true }
-
-[dev-dependencies]
-arbtest.workspace = true
-arbitrary.workspace = true
-serde_json.workspace = true
-alloy-primitives = { workspace = true, features = ["arbitrary"] }
-
-[features]
-default = ["std", "serde"]
-std = []
-arbitrary = ["std", "dep:arbitrary", "alloy-primitives/arbitrary"]
-serde = ["dep:serde", "dep:alloy-serde", "dep:serde_repr", "alloy-eips/serde", "alloy-consensus/serde", "alloy-primitives/serde"]
