--- crates/registry/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-name = "superchain-registry"
-description = "Bindings for the Superchain Registry"
-version.workspace = true
-edition.workspace = true
-license.workspace = true
-authors.workspace = true
-repository.workspace = true
-
-[dependencies]
-# Misc
-lazy_static.workspace = true
-hashbrown.workspace = true
-
-# Workspace Dependencies
-superchain-primitives = { workspace = true, default-features = false, features = ["serde"] }
-
-# Serialization
-serde.workspace = true
-serde_json.workspace = true
-
-[dev-dependencies]
-alloy-primitives.workspace = true
-
-[features]
-default = ["std"]
-std = ["superchain-primitives/std", "serde_json/std"]
