+++ crates/superchain/README.md
@@ -3,14 +3,11 @@
 The `superchain` is an optionally `no_std` crate that provides core types
 and bindings for the Superchain.
 
-It re-exports two crates:
-- [`superchain-primitives`][scp]
-- [`superchain-registry`][scr] _Only available if `serde` feature flag is enabled_
+[`superchain`][sup] is an optionally `no_std` crate, by disabling
+the `std` feature flag. By default, `std` is enabled, providing standard
+library support.
 
-[`superchain-primitives`][scp] defines core types used in the `superchain-registry`
-along with a few default values for core chains.
-
-[`superchain-registry`][scr] provides bindings to all chains in the `superchain`.
+[`serde`][s] is a required dependency.
 
 ## Usage
 
@@ -18,21 +15,21 @@ Add the following to your `Cargo.toml`.
 
 ```toml
 [dependencies]
-superchain = "0.3"
+superchain = "0.5"
 ```
 
-To make make `superchain` `no_std`, toggle `default-features` off like so.
+To make `superchain` work in a `no_std` environment, toggle `default-features` off like so.
 
 ```toml
 [dependencies]
-superchain = { version = "0.3", default-features = false }
+superchain = { version = "0.5", default-features = false }
 ```
 
 ## Example
 
-[`superchain-registry`][scr] exposes lazily defined mappings from chain id
-to chain configurations that the [`superchain`][sup] re-exports. Below
-demonstrates getting the `RollupConfig` for OP Mainnet (Chain ID `10`).
+[`superchain`][sup] exposes lazily defined mappings from chain id
+to chain configurations. Below demonstrates getting the `RollupConfig`
+for OP Mainnet (Chain ID `10`).
 
 ```rust
 use superchain::ROLLUP_CONFIGS;
@@ -54,16 +51,13 @@ println!("OP Mainnet Chain Config: {:?}", op_chain_config);
 
 ## Feature Flags
 
-- `serde`: Enables [`serde`][s] support for types and makes [`superchain-registry`][scr] types available.
 - `std`: Uses the standard library to pull in environment variables.
 
 <!-- Hyperlinks -->
 
 [sp]: ../superchain-primitives
 
+[gsr]: https://github.com/ethereum-optimism/superchain-registry
 [s]: https://crates.io/crates/serde
 [sr]: https://github.com/anton-rs/superchain
-[scr]: https://crates.io/crates/superchain-registry
 [sup]: https://crates.io/crates/superchain
-[scp]: https://crates.io/crates/superchain-primitives
-
