+++ crates/superchain/etc/chainList.json
